requirements.txt
click==6.6
psycopg2==2.6.2
python-dotenv==0.6.0
web.py==0.38
